% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateToEpiweek.R
\name{dateToEpiweek}
\alias{dateToEpiweek}
\title{This function converts date records into epidemiological week (epiweek) records.}
\usage{
dateToEpiweek(matrix, ...)
}
\arguments{
\item{matrix}{matrix}

\item{...}{further arguments passed to or from other methods}
}
\value{
matrix
}
\description{
This function converts date records into epidemiological week (epiweek) records.
}
