% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactDateToEpiweek.R
\name{exactDateToEpiweek}
\alias{exactDateToEpiweek}
\title{This function consolidates the date records into epiweek records by adding up all the incidences/samples.}
\usage{
exactDateToEpiweek(matrix, ...)
}
\arguments{
\item{matrix}{matrix}

\item{...}{further arguments passed to or from other methods}
}
\value{
matrix
}
\description{
This function consolidates the date records into epiweek records by adding up all the incidences/samples.
}
