% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expectedSampleMatrix.R
\name{expectedSampleMatrix}
\alias{expectedSampleMatrix}
\title{This function calculates and returns the expected number of genome samples by Epiweek and location, given a baseline.}
\usage{
expectedSampleMatrix(baseline, seqM, caseM, ...)
}
\arguments{
\item{baseline}{numeric}

\item{seqM}{matrix}

\item{caseM}{matrix}

\item{...}{further arguments passed to or from other methods}
}
\value{
matrix
}
\description{
This function calculates and returns the expected number of genome samples by Epiweek and location, given a baseline.
}
