% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaTableToMatrix.R
\name{metaTableToMatrix}
\alias{metaTableToMatrix}
\title{This function converts metadata from a data frame format into a matrix format.}
\usage{
metaTableToMatrix(texasSeqMeta, columnNameForLocation, columnNameForDate, ...)
}
\arguments{
\item{texasSeqMeta}{data frame}

\item{columnNameForLocation}{character string}

\item{columnNameForDate}{character string}

\item{...}{further arguments passed to or from other methods}
}
\value{
matrix
}
\description{
This function converts metadata from a data frame format into a matrix format.
}
