% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSequencingPercentage.R
\name{plotSequencingPercentage}
\alias{plotSequencingPercentage}
\title{This function plots the sequencing percentage.}
\usage{
plotSequencingPercentage(seqM, caseM, ...)
}
\arguments{
\item{seqM}{matrix}

\item{caseM}{matrix}

\item{...}{further arguments passed to or from other methods}
}
\value{
plot
}
\description{
This function plots the sequencing percentage.
}
