% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSequencingRatio.R
\name{plotSequencingRatio}
\alias{plotSequencingRatio}
\title{This function plots the sequencing ratio.}
\usage{
plotSequencingRatio(seqM, caseM, ...)
}
\arguments{
\item{seqM}{matrix}

\item{caseM}{matrix}

\item{...}{further arguments passed to or from other methods}
}
\value{
plot
}
\description{
This function plots the sequencing ratio.
}
