% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportionalSampling.R
\name{proportionalSampling}
\alias{proportionalSampling}
\title{This function performs proportional sampling and returns the genome IDs as a vector.}
\usage{
proportionalSampling(sampleM, seqMeta, ...)
}
\arguments{
\item{sampleM}{matrix}

\item{seqMeta}{dataframe}

\item{...}{further arguments passed to or from other methods}
}
\value{
vector
}
\description{
This function performs proportional sampling and returns the genome IDs as a vector.
}
