% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposedSamplingMatrix.R
\name{proposedSamplingMatrix}
\alias{proposedSamplingMatrix}
\title{This function calculates and returns the expected number of genome samples per epidemiological week and location based on a given baseline.}
\usage{
proposedSamplingMatrix(baseline, seqM, caseM, ...)
}
\arguments{
\item{baseline}{numeric}

\item{seqM}{matrix}

\item{caseM}{matrix}

\item{...}{further arguments passed to or from other methods}
}
\value{
matrix
}
\description{
This function calculates and returns the expected number of genome samples per epidemiological week and location based on a given baseline.
}
