% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{texasCase}
\alias{texasCase}
\title{SARS-CoV-2 Epi trends in TX}
\format{
\subsection{\code{texasCase} is a matrix with 26 rows and 365 columns:}{

rownames are Metro Areas in TX
colnames are date
}
}
\source{
\url{https://www.dshs.texas.gov/covid-19-coronavirus-disease-2019/texas-covid-19-data}
}
\usage{
texasCase
}
\description{
SARS-CoV-2 Epi trends in TX
}
\keyword{datasets}
