% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{texasSeqMeta}
\alias{texasSeqMeta}
\title{The metadata of SARS-CoV-2 genome samples}
\format{
\subsection{\code{texasSeqMeta}is a data frame with 24593 rows and 3 columns:}{

\describe{
\item{gisaid_epi_isl}{GISAID id}
\item{date}{sampling date}
\item{location}{sampling location}
}
}
}
\usage{
texasSeqMeta
}
\description{
The metadata of SARS-CoV-2 genome samples
}
\keyword{datasets}
