% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaTableToMatrix.R
\name{metaTableToMatrix}
\alias{metaTableToMatrix}
\title{This function transforms the metadata, which is in the form of a data frame, into a matrix.}
\usage{
metaTableToMatrix(texasSeqMeta, columnNameForLocation, columnNameForDate, ...)
}
\arguments{
\item{texasSeqMeta}{data frame}

\item{columnNameForLocation}{character string}

\item{columnNameForDate}{character string}

\item{...}{further arguments passed to or from other methods}
}
\value{
matrix
}
\description{
This function transforms the metadata, which is in the form of a data frame, into a matrix.
}
