% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{texasCase}
\alias{texasCase}
\title{Daily case counts of SARS-CoV-2 by subregion in Texas}
\format{
\subsection{\code{texasCase} is a matrix with 26 rows and 365 columns:}{

rownames are Metro Areas in TX
colnames are date (YYYY-MM-DD)
}
}
\source{
\url{https://www.dshs.texas.gov/covid-19-coronavirus-disease-2019/texas-covid-19-data}
}
\usage{
texasCase
}
\description{
Daily case counts of SARS-CoV-2 by subregion in Texas
}
\keyword{datasets}
