% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{texasSeqMeta}
\alias{texasSeqMeta}
\title{Metadata for SARS-CoV-2 Delta variant genome samples collected from Texas}
\format{
\subsection{\code{texasSeqMeta}is a data frame with 24593 rows and 3 columns:}{

\describe{
\item{gisaid_epi_isl}{GISAID id}
\item{date}{sampling date}
\item{location}{sampling location}
}
}
}
\usage{
texasSeqMeta
}
\description{
Metadata for SARS-CoV-2 Delta variant genome samples collected from Texas
}
\keyword{datasets}
